Download data on 'Low Traffic Neighbourhoods' or 'rat runs' from CycleStreets

R interface to the CycleStreets.net LTN. See ltn API docs and an article on the methods for further details: https://www.cyclestreets.org/news/2021/07/25/mapping-ltns/

ltns(bb, pat = Sys.getenv("CYCLESTREETS"))

Arguments

bb
An sf or 'bounding box' like object

pat
The API key used. By default this uses Sys.getenv("CYCLESTREETS").

Details

CycleStreets.net does not yet work worldwide. Requires the internet and a CycleStreets.net API key.

A quick way to set this is to install the usethis package and then executing the following command:

usethis::edit_r_environ()

That should open up a new file in your text editor where you can add the environment variable as follows (replace 1a... with your key for this to work):

CYCLESTREETS=1a43ed677e5e6fe9

After setting the environment variable, as outlined above, you need to restart your R session before the journey function will work.

A full list of variables (cols) available is represented by:

Examples

if (FALSE) {
bb <- "0.101131,52.195807,0.170288,52.209719"
ltn_data <- ltns(bb)
plot(ltn_data)
bb <- stplanr::routes_fast_sf
ltn_data <- ltns(bb)
plot(ltn_data)
}